--- include/graph/algorithm/example/CMakeLists.txt
@@ -1,3 +1,0 @@
-# ./example
-
-#add_subdirectory(undirected)
